init directory sync resources
