+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [7.8.0 - 2022-08-08]
 ### Added
 - create sublist for cql results
+- init resources for directory sync
 
 ## [7.7.2 - 2022-07-27]
 ### Fixed
