+++ src/main/java/de/samply/share/client/job/DirectorySyncJob.java
@@ -1,22 +1,9 @@
 package de.samply.share.client.job;
 
 import ca.uhn.fhir.context.FhirContext;
-import ca.uhn.fhir.rest.client.api.IGenericClient;
-import ca.uhn.fhir.rest.client.interceptor.LoggingInterceptor;
 import de.samply.directory_sync.Sync;
-import de.samply.directory_sync.directory.DirectoryApi;
-import de.samply.directory_sync.directory.DirectoryService;
-import de.samply.directory_sync.fhir.FhirApi;
-import de.samply.directory_sync.fhir.FhirReporting;
 import de.samply.share.client.control.ApplicationBean;
-import de.samply.share.client.model.db.enums.TargetType;
-import de.samply.share.client.model.db.tables.pojos.Credentials;
-import de.samply.share.client.util.db.CredentialsUtil;
-import io.vavr.control.Either;
-import java.io.IOException;
 import java.util.List;
-import javax.security.auth.login.CredentialNotFoundException;
-import org.apache.http.impl.client.HttpClients;
 import org.hl7.fhir.r4.model.OperationOutcome;
 import org.quartz.Job;
 import org.quartz.JobExecutionContext;
@@ -41,12 +28,8 @@ public class DirectorySyncJob implements Job {
     }
   }
 
-  private void syncWithDirectory() throws CredentialNotFoundException, IOException {
-    DirectoryApi directoryApi = createDirectoryApi().get();
-    DirectoryService directoryService = new DirectoryService(directoryApi);
-    FhirApi fhirApi = createFhirApi();
-    FhirReporting fhirReporting = new FhirReporting(ctx, fhirApi);
-    Sync sync = new Sync(fhirApi, fhirReporting, directoryApi, directoryService);
+  private void syncWithDirectory() {
+    Sync sync = ApplicationBean.getSync();
     List<OperationOutcome> operationOutcomes = sync.syncCollectionSizesToDirectory();
     for (OperationOutcome oo : operationOutcomes) {
       logger.debug(ctx.newJsonParser().encodeResourceToString(oo));
@@ -57,23 +40,4 @@ public class DirectorySyncJob implements Job {
     }
   }
 
-  private Either<OperationOutcome, DirectoryApi> createDirectoryApi()
-      throws CredentialNotFoundException, IOException {
-    List<Credentials> credentialsList = CredentialsUtil
-        .getCredentialsByTarget(TargetType.TT_DIRECTORY);
-    if (credentialsList.size() < 1) {
-      throw new CredentialNotFoundException("No directory credentials found");
-    }
-    Credentials credentials = credentialsList.get(0);
-    return DirectoryApi.createWithLogin(HttpClients.createDefault(),
-        ApplicationBean.getUrlsForDirectory().getDirectoryUrl(), credentials.getUsername(),
-        credentials.getPasscode());
-  }
-
-  private FhirApi createFhirApi() {
-    IGenericClient client = ctx
-        .newRestfulGenericClient(ApplicationBean.getUrlsForDirectory().getLdmUrl());
-    client.registerInterceptor(new LoggingInterceptor(true));
-    return new FhirApi(client);
-  }
 }
